@@ -11,19 +11,23 @@
     "@react-navigation/bottom-tabs": "^6.0.9",
     "@react-navigation/native": "^6.0.6",
     "@react-navigation/stack": "^6.0.11",
-    "expo": "^42.0.5",
-    "expo-status-bar": "~1.0.4",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-42.0.0.tar.gz",
+    "expo": "^44.0.0",
+    "expo-status-bar": "~1.2.0",
+    "firebase": "^9.4.1",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "react-native": "0.64.3",
     "react-native-dropdown-picker": "^5.2.3",
     "react-native-elements": "^3.4.2",
-    "react-native-gesture-handler": "^1.10.3",
-    "react-native-maps": "0.28.0",
-    "react-native-web": "^0.17.5"
+    "react-native-firebase": "^5.6.0",
+    "react-native-gesture-handler": "~2.1.0",
+    "react-native-maps": "0.29.4",
+    "react-native-safe-area-context": "^3.3.2",
+    "react-native-uuid": "^2.0.1",
+    "react-native-web": "0.17.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.16.0"
+    "@babel/core": "^7.12.9"
   },
   "private": true
 }
